Use prepended modules instead of modifying the target class
If the method being overidden by this gem is also overidden by a prepended
module, the result ends up being infinite recursion as `old_method` will always
be pointing to the method added by the module, and `super` there will point to
the addition from `StatsD::Instrument`. Normally I would just prepend a single
anonymous module every time this was called, but I've attempted to preserve the
"has this already been instrumented" checks and the ability to remove
instrumentations later. Instead we keep a hash of various modules, and
define/undefine the method being overidden. It's impossible to remove a module
after it's been added, but when `statsd_remove_measure` or friends are called
we simply undefine the methdos on the module leaving us with an empty module
doing nothing.

This does mean that we'll end up with a lot of anonymous modules in the
ancestors chain, but there's really no way to avoid that. I have given them a
more descriptive name at least in order to aid debugging if that occurs

Fixes #91.