Fix bug in escaping newlines for events with Datadog Backend
Ran into this issue where we were sending events with newlines that were
NOT getting properly escaped.

This was due to a few compounding issues:

1. `'\n'` is NOT the same as `"\n"` in Ruby. Single quotes do NOT treat
   `\` special so that first string is 2 chars. While the double quoted
   string takes `\n` to mean the newline character
1. `tr` only ever replaces single characters. The usage of `tr` does not
   appear to be correct. Since the intended behavior is to take a single
   char `\n` (the new line character) and replace it with TWO chars, `\`
   followed by an `n`, or `"\\n"` in Ruby. Since `tr` can only replace
   single chars it does not seem well suited for this problem

This replaces the usage of `tr` with `gsub` which has the right
behavior.

It also updates the tests to use double quotes, since their use of
single quotes appears to be hiding this bug.