Refactor assert_statsd_calls to flunk if an exception occurs inside the block provided to the assertion method.
This is a more straightforward way of dealing with exception that our previous approach. We tell people they should nest `assert_raises` inside the statsD assertion block.

This is a backwards incompatible change though - it may cause some tests to fail that previously were succeeding. However, the fix for those is easy: change the nesting order.