Load strict mode if STATSD_STRICT_MODE is set.
We now only use this environment variable for our own test suite, but it is very useful for applications that use this gem as well to quickly run tests in strict mode.

Updating the Gemfile is harder, and you have to be careful to only do it in test mode, and not to accidentally enable it in production.