Always sample datagrams handled by a NullSink.
NullSink#sample? currently always returns `false`, which prevents us from generating a datagram unnecessarily. This seems like a simple performance boost, but comes with a price.

The NullSink is primarily used in the test environment. In this environment, it can make a lot of sense to use a custom DatagramBuilder that enforces certain rules about the metrics you are emitting. But because sample? Returns false, this datagram builder is never called (except when capturing datagrams from the NullSink).

Because high performance is not really the main use case for the test environment, it makes more sense to always return true, so we always go through the datagram builder.