Remove outdated StatsD test.
StatsD.measure and StatsD.distribution will not raise an ArgumentError if no value or block is provided. We assume people will know how to use the API, so we don’t incur a penalty of this check every time we emit a metric.

In strict mode, this check is still active, which will help you ensure in development and CI that you are using the API correctly.