Wake up the dispatch thread once the buffer is past a threshold
The current sleep loop can lead to large memory usage if lots of
events are queued rapidly. In such scenario the buffer can grow
very large until the threads wake up and flush it.

To avoid this, after enqueing a message, if the buffer reached
a threshold we explictly wake up the dispatch thread to trigger
a flush.