BatchedUDPSink: block when the queue is full
If an application emit so much metrics that the dispatcher thread
can't keep up, the buffer may grow unbounded.

This is particularly aggravated by the the thread scheduler that
tries not to give the dispatcher thread a disproportionate slice of
CPU.

To counter act that, this commit introduce a max_queue_size. Once
reached emitter thread will wait for the buffer to be emptied.

This means that applications emitted too many events will be slowed
down like they would with the regular UDP Sink. However the only
alternative is to drop events.

Additionally, after flushing the buffer, the dispatcher thread
now immediately check wether the threshold was reached again.
This should improve throughput, as threads may have filled the
buffer while the thread was doing IOs.

We should however explore other ways to increase the UDP throughput.