Allow user to assert different tags using RSpec
Previously the matcher filtered out metrics based on type and name to
test against expectation. So, when the user tried to assert different
different tags, the matcher would fail.
With this change, the matcher will filter out metrics also based on tags.
Allowing the user to use RSpec `and` expectation.