Stop variable shadowing from breaking dynamic tags
Re-assigning the `tags` variable defined outside a block inside the
block overrides that value for subsequent calls. That caused dynamic
tags to be evaluated once and then be fixed to that first value —
rather than being dynamically evaluated on every invocation. By using a
unique variable name, this is prevented from happening.