introducing periodical flushes
this is also common practice and I wanted to avoid doing that, but seems like we dont have much way to go around it. Otherwise the samples will never make out of the client, only on shutdown.

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>

remove unused sample_rate arg for increment in aggregator

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>

wire flush in client

Signed-off-by: Pedro Tanaka <pedro.tanaka@shopify.com>