Refactor StatsD::Client#emit to take arguments
Prior to this commit, `statsd-instrument` eagerly generated serialized
StatsD lines when calling methods like `Client#increment`, and then
passed those strings directly to the configured sink via the `emit`
method. However, while there are benefits to eager-serialization, it
means that anything which wishes to process StatsD calls as part of
statsd-instrument must deal entirely with strings.

In order to more easily experiment with things that modify or alter
statsd-instrument's behavior, we refactor this to take specific, named
parameters. This allows interested parties to override the `Client#emit`
method and intercept StatsD traffic _before_ serialization, in order to
do whatever they want (such as aggregate, or pass to a third-party sink
implementation, etc).