Add basic, single-message messagepack datagram builder
This commit will serialize a statsd datagram as a custom messagepack
extension type (and many details around this type are TBD; and not all
datagram types are actually supported).

For now, it'll just serialize one datagram into a single messagepack
message. We are setting `values` as an array (in the hopes of value
packing), but this commit doesn't implement that.

We also can't reliably serialize multiple of these calls into a batched
packet, so you'd want to use the regular `UDPSink` with this:

```
require 'statsd/instrument'
client = StatsD.singleton_client.clone_with_options(
   sink: StatsD::Instrument::UDPSink.new(host, port),
   datagram_builder_class: StatsD::Instrument::MessagePackDatagramBuilder,
)

client.increment("foo", 1) # or whatever
```

This is just something of a starting point, and this particular commit
will probably not merge into main as-is.