[wip] Batched / Unbatched SysVMQ Sink
Thought it would be interesting to try using SysVMQ as a sink rather
than UDP - and initial benchmarks show basically the following ranking
(with 1 being the fastest):

1. Batched SysVMQ
2. Batched UDP
3. Unbatched SysvMQ
4. Unbatched UDP

This is perhaps not terribly surprising, but it did surprise me a
*little* bit to be honest.

The main downside to SysV MQ of course is that it is not a
fire-and-forget thing like UDP. If something isn't listening, then the
queue fills up and in *this* implementation, that would mean we'd block
indefinitely (it's possible to do non-blocking, drop-it-on-the-floor
sends with a different option to the `@mq.send` function).

Notably, to run the benchmark you'll need to run either the
`benchmark/sysvmq-receiver` or compile+run `benchmark/sysvmq-receiver.c`
_before_ you start the benchmark script - otherwise nothing will be
listening for the datagrams. In my testing, the ruby receiver only added
a miniscule amount of overhead for benchmarks, so running either one is
fine.